with sanity check for 1 parameter

+++ bulkonvert.sh
@@ -7,15 +7,24 @@
 # Sanity checks ?
 # ...
 
+if [ $# != 1 ] ; 
+  then
+    echo ""
+    echo "bad parameter count"
+    echo "you could write ./bulkonvert.sh with one parameter only"
+    echo ""
+    exit 2
+else
 # Main code
 
-ls -1 $1*.avi > videofiles
+  ls -1 $1*.avi > videofiles
 
-while videofile
-do
+  while videofile
+  do
 	lengthname=${#videofile}
 	namefile=${videofile;0;$lengthname-4}
-	mencoder $videofile -o namefile.mp4 -oac mp3lame -ovc lavc -of mp4
+  	mencoder $videofile -o namefile.mp4 -oac mp3lame -ovc lavc -of mp4
 
-done < videofiles
-rm videofiles
+  done < videofiles
+  rm videofiles
+fi
